+++ shared/Tracer/lib/TransactionTracker.cs
@@ -11,10 +11,10 @@ namespace Tracer
         public TraceProject Project { get; set; }
 
         //Todo change partion back to uint32
-        protected UInt64 partitionOffset_ { get; set; } = 0;
+        protected UInt32 partitionOffset_ { get; set; } = 0;
         protected List<TraceTransactionLog> logs_ { get; set; }
 
-        public TransactionTracker(UInt64 partitionSize)
+        public TransactionTracker(UInt32 partitionSize)
         {
             Project = new TraceProject()
             {
@@ -28,21 +28,21 @@ namespace Tracer
 
         //For loading existing transactions
         //Logs should only be after the offset
-        public TransactionTracker(TraceProject project, List<TraceTransactionLog> transactionLogs, UInt64 partitionOffset)
+        public TransactionTracker(TraceProject project, List<TraceTransactionLog> transactionLogs, UInt32 partitionOffset)
         {
             Project = project;
             partitionOffset_ = partitionOffset;
             logs_ = transactionLogs;
         }
 
-        public TraceTransactionLog GetTransactionLogByTimeOffset(UInt64 timeOffset)
+        public TraceTransactionLog GetTransactionLogByTimeOffset(UInt32 timeOffset)
         {
             TraceTransactionLog transactionLog = null;
             var partition = Project.PartitionFromOffsetBottom(timeOffset);
-            while (partition > ((UInt64)logs_.Count))
+            while (partition > (logs_.Count))
             {
                 transactionLog = new TraceTransactionLog();
-                transactionLog.Partition = partitionOffset_ + (UInt64)(logs_.Count - 1);
+                transactionLog.Partition = partitionOffset_ + (UInt32)(logs_.Count - 1);
                 logs_.Add(transactionLog);
             }
 
@@ -61,7 +61,7 @@ namespace Tracer
             Project.Duration += transaction.TimeOffsetMs;
         }
 
-        public void CreateFile(UInt64 timeOffset, string file_path)
+        public void CreateFile(UInt32 timeOffset, string file_path)
         {
             var transaction = new TraceTransaction();
             transaction.Type = TraceTransaction.Types.TraceTransactionType.CreateFile;
@@ -71,7 +71,7 @@ namespace Tracer
             transaction.CreateFile = data;
         }
 
-        public void DeleteFile(UInt64 timeOffset, string file_path)
+        public void DeleteFile(UInt32 timeOffset, string file_path)
         {
             var transaction = new TraceTransaction();
             transaction.Type = TraceTransaction.Types.TraceTransactionType.CreateFile;
@@ -81,7 +81,7 @@ namespace Tracer
             transaction.DeleteFile = data;
         }
 
-        public void InsertFile(UInt64 timeOffset, string file_path, UInt64 line, UInt64 offset, string insertData)
+        public void InsertFile(UInt32 timeOffset, string file_path, UInt32 line, UInt32 offset, string insertData)
         {
             var transaction = new TraceTransaction();
             transaction.Type = TraceTransaction.Types.TraceTransactionType.CreateFile;
@@ -94,7 +94,7 @@ namespace Tracer
             transaction.InsertFile = data;
         }
 
-        public void EraseFile(UInt64 timeOffset, string file_path, UInt64 line, UInt64 offsetStart, UInt64 offsetEnd)
+        public void EraseFile(UInt32 timeOffset, string file_path, UInt32 line, UInt32 offsetStart, UInt32 offsetEnd)
         {
             var transaction = new TraceTransaction();
             transaction.Type = TraceTransaction.Types.TraceTransactionType.CreateFile;
